Fix LGTM warnings (#50)
This commit fixes LGTM errors and warnings.

The file netnscontroller.py is removed in this process. The file is
seemingly a debug copy of netns_base.py. This class expects serial
number in making netns calls, but the main method in this file simply
uses netns_base.py signatures that don't supply serial numbers.

Background about the changes in Watchable class: LGTM emits warning
for self.value assignments because of Python descriptor's
properties. __set__ method's self refers to the class, so using
self.value will supposedly set the value for all instances of the
class using the descriptor method.